// doc-backend/src/shared/enum/common.enum.ts
export enum Gender{
    M = 'male',
    F = 'female'
}
export enum UserType {
    SUPERADMIN = 'super-admin',
    OWNER = 'admin',
    DOCTOR = 'doctor',
    OPERATOR = 'operator',
    MANAGER = 'manager',
    ACCOUNTANT = 'accountant'
}
export enum Subscriptions{
    MONTHLY  = 'monthly',
    QUARTERLY = 'quarterly ',
    BIANNUALY = 'bi-annualy',
    ANNUALY = 'annualy'
}
export enum RelationShip{
    FATHER = 'father',
    MOTHER = 'mother',
    BROTHER = 'brother',
    SISTER = 'sister',
    UNCLE = 'uncle',
    AUTN = 'aunt',
    SON = 'son',
    DAUGHTER = 'daughter',
    NEPHEW = 'newphew',
    OTHER = 'other',
    HUSBAND = 'husband',
    WIFE = 'wife'
}